requirements.txt
protobuf==3.6.1
python-memcached==1.59